+++ src/Lakona.Tool/Domain/ClientEngine.cs
@@ -3,7 +3,6 @@ namespace Lakona.Tool.Domain;
 internal enum ClientEngine
 {
     Unity,
-    UnityCn,
     Tuanjie,
     Godot,
     Console